Move ShellScript out of the workflow module
