+++ crates/core/tedge_api/src/workflow/script.rs
@@ -1,71 +1,10 @@
 use crate::workflow::GenericStateUpdate;
 use crate::workflow::ScriptDefinitionError;
-use serde::de::Error;
-use serde::Deserialize;
-use serde::Deserializer;
-use serde::Serialize;
-use serde::Serializer;
 use serde_json::Value;
 use std::cmp::max;
 use std::fmt::Display;
-use std::fmt::Formatter;
 use std::os::unix::prelude::ExitStatusExt;
-use std::str::FromStr;
 use std::time::Duration;
-
-/// A parsed Unix command line
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct ShellScript {
-    pub command: String,
-    pub args: Vec<String>,
-}
-
-/// Deserialize an Unix command line
-impl<'de> Deserialize<'de> for ShellScript {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let command_line = String::deserialize(deserializer)?;
-        command_line.parse().map_err(Error::custom)
-    }
-}
-
-impl FromStr for ShellScript {
-    type Err = String;
-
-    fn from_str(command_line: &str) -> Result<Self, Self::Err> {
-        let mut args = shell_words::split(command_line)
-            .map_err(|err| format!("invalid script: {command_line}: {err}"))?;
-        if args.is_empty() {
-            Err("invalid script: empty".to_string())
-        } else {
-            let script = args.remove(0);
-            Ok(ShellScript {
-                command: script,
-                args,
-            })
-        }
-    }
-}
-
-/// Serialize an Unix command line, using appropriate quotes
-impl Serialize for ShellScript {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        self.to_string().serialize(serializer)
-    }
-}
-
-impl Display for ShellScript {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        let mut args = vec![self.command.clone()];
-        args.append(&mut self.args.clone());
-        f.write_str(&shell_words::join(args))
-    }
-}
 
 /// Define how to interpret the exit code of a script as the next state for a command
 #[derive(Clone, Debug, Default, Eq, PartialEq)]
@@ -396,23 +335,12 @@ impl Default for DefaultHandlers {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::script::ShellScript;
     use crate::workflow::OperationAction;
     use crate::workflow::OperationWorkflow;
     use serde_json::json;
     use std::process::Command;
-
-    #[test]
-    fn script_parse_and_display() {
-        let script: ShellScript = "sh -c 'sleep 10'".parse().unwrap();
-        assert_eq!(
-            script,
-            ShellScript {
-                command: "sh".to_string(),
-                args: vec!["-c".to_string(), "sleep 10".to_string()]
-            }
-        );
-        assert_eq!(format!("{script}"), "sh -c 'sleep 10'");
-    }
+    use std::str::FromStr;
 
     #[test]
     fn successful_exit_code_determines_next_state() {
