+++ crates/core/tedge_api/src/workflow/toml_config.rs
@@ -1,4 +1,5 @@
 use crate::mqtt_topics::OperationType;
+use crate::script::ShellScript;
 use crate::workflow::AwaitHandlers;
 use crate::workflow::DefaultHandlers;
 use crate::workflow::ExecHandlers;
@@ -9,7 +10,6 @@ use crate::workflow::IterateHandlers;
 use crate::workflow::OperationAction;
 use crate::workflow::OperationWorkflow;
 use crate::workflow::ScriptDefinitionError;
-use crate::workflow::ShellScript;
 use crate::workflow::WorkflowDefinitionError;
 use serde::de::Error;
 use serde::Deserialize;
